Pull all event conversion back into MessagingService
